requirements.txt
requests
PyYAML
Flask
tweepy